requirements.txt
Flask>=2.0.1
openai>=1.20.0
python-dotenv>=0.19.0
pandas>=2.0.0
numpy>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.4.3
plotly>=5.3.1
Flask-WTF>=1.0.0 